server_api/requirements.txt
backcall==0.1.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
decorator==4.4.0
Django==1.11.23
django-cors-headers==3.1.0
django-filter==2.2.0
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
idna==2.8
ipython==7.8.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.15.1
Jinja2==2.10.1
MarkupSafe==1.1.1
mysqlclient==1.4.4
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
PyJWT==1.7.1
pytz==2019.2
requests==2.22.0
six==1.12.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.25.3
wcwidth==0.1.7
